Move HTML attribute observer controller to new controllers directory
