+++ controllers/attributeObserver/htmlAttributeObserverController.js
