+++ controllers/attributeObserver/test/htmlAttributeObserverController.test.js
